# server/src/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Digital-E Shop API",
        "description": "REST API reference for the Digital-E e-commerce backend. Covers catalog browsing, authentication, cart, orders, wishlist, reviews, address book, notifications, inventory, promotions, analytics, and admin operations.",
        "version": "1.0.0",
        "contact": {
            "name": "Digital-E",
            "url": "https://github.com/anomalyco/digital-e-shop"
        },
        "license": {
            "name": "ISC"
        }
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Local development"
        },
        {
            "url": "https://e-commerce-express-server-app.vercel.app",
            "description": "Production"
        }
    ],
...
    "components": {
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "msg": { "type": "string", "description": "Human-readable error message" },
                    "error": { "type": "string", "description": "Machine-readable error code" }
                }
            },
...
            "Product": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer" },
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "category": { "type": "string" },
                    "brand": { "type": "string" },
                    "price": { "type": "number" },
                    "sale_price": { "type": "number", "nullable": true },
                    "stock": { "type": "integer" },
                    "main_image": { "type": "string", "nullable": true },
                    "specifications": { "type": "string", "nullable": true },
                    "rating": { "type": "number" },
                    "reviews": { "type": "integer" }
                }
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "accessToken",
                "description": "JWT access token stored in the accessToken HTTP-only cookie."
            }
        }
    },
    "paths": {
...
        "/api/products/search": {
            "get": {
                "tags": ["Products"],
                "summary": "Search products",
                "description": "Full-text search across product name, brand, category, and description.",
                "parameters": [
                    { "name": "q", "in": "query", "required": true, "schema": { "type": "string", "minLength": 2 } },
                    { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 20, "default": 6 } }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/recommendations/{uid}": {
            "parameters": [
                { "name": "uid", "in": "path", "required": true, "schema": { "type": "string" } },
                { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 24, "default": 12 } }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Get personalized product recommendations for a user",
                "description": "Recommendation engine based on purchase history, wishlist, cart activity, and brand affinity.",
                "responses": {
                    "200": {
                        "description": "Recommended products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "products": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
                                    }
                                }
                            }
                        }
                    },
                    "500": { "$ref": "#/components/responses/Error" }
                }
            }
        },
        "/api/products/relevant/{pid}": {
            "parameters": [
                { "name": "pid", "in": "path", "required": true, "schema": { "type": "integer" } }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Get products relevant to a given product",
                "description": "Returns up to 8 related products based on shared category, brand, co-purchase history, ratings, and price proximity.",
                "responses": {
                    "200": {
                        "description": "Relevant products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "relevantProducts": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
                                    }
                                }
                            }
                        }
                    },
                    "400": { "$ref": "#/components/responses/Error" }
                }
            }
        },
        "/api/products/images/{filename}": {
            "parameters": [
                { "name": "filename", "in": "path", "required": true, "schema": { "type": "string" } }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Serve a product image",
                "description": "Returns a JPEG image from the local uploads directory.",
                "responses": {
                    "200": {
                        "description": "Image bytes",
                        "content": { "image/jpeg": { "schema": { "type": "string", "format": "binary" } } }
                    },
                    "400": { "description": "Invalid file path" }
                }
            }
        },
...
        "/api/products/add": {
            "post": {
                "tags": ["Products"],
                "summary": "Add a new product (admin)",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": ["name", "category", "brand", "price", "inventory"],
                                "properties": {
                                    "name": { "type": "string" },
                                    "description": { "type": "string" },
                                    "category": { "type": "string" },
                                    "brand": { "type": "string" },
                                    "specifications": { "type": "string" },
                                    "price": { "type": "number" },
                                    "inventory": { "type": "integer" },
                                    "image": { "type": "string", "format": "binary" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "$ref": "#/components/responses/Message" },
                    "400": { "$ref": "#/components/responses/Error" },
                    "403": { "$ref": "#/components/responses/Error" }
                }
            }
        },
        "/api/products/{id}/inventory": {
            "parameters": [{ "$ref": "#/components/parameters/PathId" }],
            "put": {
                "tags": ["Products"],
                "summary": "Quick restock (admin)",
                "description": "Updates the product's stock and records an inventory movement.",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["stock"],
                                "properties": { "stock": { "type": "integer", "minimum": 0 } }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "product": { "$ref": "#/components/schemas/Product" }
                                    }
                                }
                            }
                        }
                    },
                    "400": { "$ref": "#/components/responses/Error" },
                    "404": { "$ref": "#/components/responses/Error" }
                }
            }
        },
        "/api/products/{id}": {
            "parameters": [{ "$ref": "#/components/parameters/PathId" }],
            "put": {
                "tags": ["Products"],
                "summary": "Update a product (admin)",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "description": { "type": "string" },
                                    "category": { "type": "string" },
                                    "brand": { "type": "string" },
                                    "specifications": { "type": "string" },
                                    "price": { "type": "number" },
                                    "salePrice": { "type": ["number", "null", ""] },
                                    "stock": { "type": "integer" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "$ref": "#/components/responses/Message" },
                    "400": { "$ref": "#/components/responses/Error" }
                }
            }
        },
        "/api/products": {
            "delete": {
                "tags": ["Products"],
                "summary": "Delete products (admin)",
                "description": "Soft-deletes products by setting stock to -1.",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "pid": { "oneOf": [{ "type": "integer" }, { "type": "array", "items": { "type": "integer" } }] }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "$ref": "#/components/responses/Message" }
                }
            }
        },
...
    },
    "components": {
...
        "parameters": {
            "PathId": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
                "description": "Resource ID (user, product, etc.)"
            },
...
        },
        "responses": {
            "Message": {
                "description": "Standard message response",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": { "msg": { "type": "string" } }
                        }
                    }
                }
            },
            "Error": {
                "description": "Error response",
                "content": {
                    "application/json": {
                        "schema": { "$ref": "#/components/schemas/Error" }
                    }
                }
            }
        }
    }
}